% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{add_provider_specialty_filter}
\alias{add_provider_specialty_filter}
\title{Add ProviderSpecialty filter support to a domain builder}
\usage{
add_provider_specialty_filter(criteria, cdm_schema, alias_prefix)
}
\value{
list with select_col (to add to inner SELECT), join_sql, where_parts
}
\description{
Add ProviderSpecialty filter support to a domain builder
}
