% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortUtilities.R
\name{appendCohortAttributes}
\alias{appendCohortAttributes}
\title{Create or update the GeneratedCohortSet attributes to a cohort object}
\usage{
appendCohortAttributes(
  cohort,
  name = sprintf("dbplyr_\%03i", getOption("dbplyr_table_name", 0)),
  attritionReason = "Qualifying initial records",
  cohortSet = attr(cohort, "cohort_set"),
  cdm = attr(cohort, "cdm_reference")
)
}
\arguments{
\item{cohort}{Cohort in the cdm.}

\item{name}{Name of the cohort table.}

\item{attritionReason}{The reason for attrition as a character string.}

\item{cohortSet}{tbl to update the cohort_set attribute.}

\item{cdm}{A cdm_reference object. If not provided the one linked to cohort
will be used.}
}
\value{
The cohort object with the attributes created or updated.
}
\description{
Create or update the GeneratedCohortSet attributes to a cohort object
}
\examples{
\dontrun{
library(CDMConnector)
library(dplyr)

con <- DBI::dbConnect(duckdb::duckdb(), eunomia_dir())
cdm <- cdm_from_con(con = con, cdm_schema = "main", write_schema = "main")
cdm <- generateConceptCohortSet(
  cdm = cdm, conceptSet = list(pharyngitis = 4112343), name = "new_cohort"
)

cohortSet(cdm$new_cohort)
cohortCount(cdm$new_cohort)
cohortAttrition(cdm$new_cohort)

cdm$new_cohort <- cdm$new_cohort \%>\%
  filter(cohort_start_date >= as.Date("2010-01-01"))

cdm$new_cohort <- appendCohortAttributes(
  cohort = cdm$new_cohort, reason = "Only events after 2010"
)

cohortSet(cdm$new_cohort)
cohortCount(cdm$new_cohort)
cohortAttrition(cdm$new_cohort)
}

}
