% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{assert_cdm_column_names}
\alias{assert_cdm_column_names}
\title{Check column names of tables in a cdm reference}
\usage{
assert_cdm_column_names(cdm)
}
\arguments{
\item{cdm}{A cdm reference object}
}
\description{
Throws an error if any column names are missing or don't match the v5.4 CDM spec.
}
