% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkCDMConnector.R
\name{benchmarkCDMConnector}
\alias{benchmarkCDMConnector}
\title{Run benchmark of tasks using CDMConnector}
\usage{
benchmarkCDMConnector(cdm)
}
\arguments{
\item{cdm}{A CDM reference object}
}
\value{
a tibble with time taken for different analyses
}
\description{
Run benchmark of tasks using CDMConnector
}
