% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_cohort_query_internal}
\alias{build_cohort_query_internal}
\title{Build cohort query from cohort expression}
\usage{
build_cohort_query_internal(cohort, options = list())
}
\arguments{
\item{cohort}{List - cohort expression (from cohort_expression_from_json)}

\item{options}{List - build options with cdm_schema, vocabulary_schema, target_table, cohort_id, etc.}
}
\value{
List with \code{full_sql} (complete script), \code{tail_sql}, \code{primary_events_sql}, \code{codeset_sql}, and other fragments for batch assembly.
}
\description{
Build cohort query from cohort expression
}
