% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_concept_set_expression_query}
\alias{build_concept_set_expression_query}
\title{Build concept set expression query}
\usage{
build_concept_set_expression_query(
  expression,
  vocabulary_schema = "@vocabulary_database_schema"
)
}
\arguments{
\item{expression}{List - ConceptSetExpression with items}

\item{vocabulary_schema}{Schema for vocabulary tables (default: @vocabulary_database_schema)}
}
\value{
Character SQL query
}
\description{
Generates SQL for a concept set expression (include/exclude, descendants, mapped).
}
