% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_condition_occurrence_sql}
\alias{build_condition_occurrence_sql}
\title{Build Condition Occurrence SQL}
\usage{
build_condition_occurrence_sql(criteria, cdm_schema = "@cdm_database_schema")
}
\arguments{
\item{criteria}{Criteria list.}

\item{cdm_schema}{CDM schema name.}
}
\description{
Build Condition Occurrence SQL
}
