% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_date_range_clause}
\alias{build_date_range_clause}
\title{Build date range clause}
\usage{
build_date_range_clause(sql_expression, date_range)
}
\arguments{
\item{sql_expression}{SQL column expression}

\item{date_range}{List with op, value, extent}
}
\value{
SQL WHERE clause or NULL
}
\description{
Build date range clause
}
