% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_death_sql}
\alias{build_death_sql}
\title{Build Death SQL}
\usage{
build_death_sql(criteria, cdm_schema = "@cdm_database_schema")
}
\arguments{
\item{criteria}{Criteria list.}

\item{cdm_schema}{CDM schema name.}
}
\description{
Build Death SQL
}
