% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_measurement_sql}
\alias{build_measurement_sql}
\title{Build Measurement SQL}
\usage{
build_measurement_sql(criteria, cdm_schema = "@cdm_database_schema")
}
\arguments{
\item{criteria}{Criteria list.}

\item{cdm_schema}{CDM schema name.}
}
\description{
Build Measurement SQL
}
