% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_numeric_range_clause}
\alias{build_numeric_range_clause}
\title{Build numeric range clause}
\usage{
build_numeric_range_clause(sql_expression, numeric_range, format = NULL)
}
\arguments{
\item{sql_expression}{SQL column expression}

\item{numeric_range}{List with op, value, extent}

\item{format}{Optional format for floats}
}
\value{
SQL WHERE clause or NULL
}
\description{
Build numeric range clause
}
