% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_observation_period_sql}
\alias{build_observation_period_sql}
\title{Build Observation Period SQL}
\usage{
build_observation_period_sql(criteria, cdm_schema = "@cdm_database_schema")
}
\description{
Build Observation Period SQL
}
