% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_observation_sql}
\alias{build_observation_sql}
\title{Build Observation SQL}
\usage{
build_observation_sql(criteria, cdm_schema = "@cdm_database_schema")
}
\arguments{
\item{criteria}{Criteria list.}

\item{cdm_schema}{CDM schema name.}
}
\description{
Build Observation SQL
}
