% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_payer_plan_period_sql}
\alias{build_payer_plan_period_sql}
\title{Build Payer Plan Period SQL (simplified)}
\usage{
build_payer_plan_period_sql(criteria, cdm_schema = "@cdm_database_schema")
}
\description{
Build Payer Plan Period SQL (simplified)
}
