% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_procedure_occurrence_sql}
\alias{build_procedure_occurrence_sql}
\title{Build Procedure Occurrence SQL}
\usage{
build_procedure_occurrence_sql(criteria, cdm_schema = "@cdm_database_schema")
}
\description{
Build Procedure Occurrence SQL
}
