% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_text_filter_clause}
\alias{build_text_filter_clause}
\title{Build text filter clause}
\usage{
build_text_filter_clause(text_filter, column_name)
}
\arguments{
\item{text_filter}{List with text, op or character}

\item{column_name}{SQL column name}
}
\value{
SQL WHERE clause or NULL
}
\description{
Build text filter clause
}
