% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{buildCohortQuery}
\alias{buildCohortQuery}
\title{Generate SQL from cohort expression}
\usage{
buildCohortQuery(cohort, options = NULL)
}
\arguments{
\item{cohort}{List or character - cohort expression (from cohortExpressionFromJson) or JSON string}

\item{options}{List - build options (optional)}
}
\value{
Character - SQL query string
}
\description{
Generate SQL from cohort expression
}
