% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{builder_utils_section}
\alias{builder_utils_section}
\title{Builder utilities for SQL query generation}
\description{
Utility functions for building SQL from cohort criteria.
Mirrors Java CIRCE-BE BuilderUtils.
}
\keyword{internal}
