% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm_reference_methods.R
\name{$.cdm_reference}
\alias{$.cdm_reference}
\title{Subset a cdm reference object}
\usage{
\method{$}{cdm_reference}(x, name)
}
\arguments{
\item{x}{A cdm reference}

\item{name}{The name of the table to extract from the cdm object}
}
\value{
A single cdm table reference
}
\description{
Subset a cdm reference object
}
