% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{cdm_from_con}
\alias{cdm_from_con}
\title{Create a CDM reference reference to a database connection}
\usage{
cdm_from_con(
  con,
  cdm_schema = NULL,
  select = tbl_group("default"),
  write_schema = NULL
)
}
\arguments{
\item{con}{A DBI database connection to a database where an OMOP CDM v5.4 instance is located.}

\item{select}{Which tables should be included?
"all" for all CDM tables
"vocab" for just the vocabulary tables
"clinical" for just the clinical tables
.}

\item{schema}{The schema where the vocab tables are located. Defaults to NULL.}
}
\value{
A list of dplyr database table references pointing to CDM tables
}
\description{
Create a CDM reference reference to a database connection
}
