% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{cdm_from_files}
\alias{cdm_from_files}
\title{Create a CDM reference from a folder containing parquet, csv, or feather files}
\usage{
cdm_from_files(
  path,
  cdm_tables = tbl_group("default"),
  format = "auto",
  as_data_frame = TRUE
)
}
\arguments{
\item{path}{A folder where an OMOP CDM v5.4 instance is located.}

\item{cdm_tables}{Which tables should be included? Supports tidyselect and custom selection groups.
\itemize{
\item{tbl_group("all")}{all CDM tables}
\item{tbl_group("vocab")}{the CDM vocabulary tables}
\item{tbl_group("clinical")}{the clinical CDM tables}
}}

\item{cdm_schema}{The schema where the OMOP CDM tables are located. Defaults to NULL.}

\item{write_schema}{An optional schema in the CDM database that the user has write access to.}
}
\value{
A list of dplyr database table references pointing to CDM tables
}
\description{
Create a CDM reference from a folder containing parquet, csv, or feather files
}
