% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{cdmCon}
\alias{cdmCon}
\title{Get underlying database connection}
\usage{
cdmCon(cdm)
}
\arguments{
\item{cdm}{A cdm reference object created by \code{cdmFromCon}}
}
\value{
A reference to the database containing tables in the cdm reference
}
\description{
Get underlying database connection
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomiaDir())

cdm <- cdmFromCon(con = con, cdmName = "Eunomia",
                  cdmSchema =  "main", writeSchema = "main")

cdmCon(cdm)

DBI::dbDisconnect(con)
}
}
