% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{cdmWriteSchema}
\alias{cdmWriteSchema}
\title{Get cdm write schema}
\usage{
cdmWriteSchema(cdm)
}
\arguments{
\item{cdm}{A cdm reference object created by \code{cdm_from_con}}
}
\value{
The database write schema
}
\description{
Get cdm write schema
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())

cdm <- cdm_from_con(con = con, cdm_name = "Eunomia",
                    cdm_schema =  "main", write_schema = "main")

cdmWriteSchema(cdm)

DBI::dbDisconnect(con)
}
}
