% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_cohort_transformations.R
\name{cohort_erafy}
\alias{cohort_erafy}
\alias{cohortErafy}
\title{Collapse cohort records within a certain number of days}
\usage{
cohort_erafy(x, gap)

cohortErafy(x, gap)
}
\arguments{
\item{x}{A generated cohort set}

\item{gap}{When two cohort records are {gap} days apart or less the periods will be
collapsed into a single record}
}
\value{
A lazy query on a generated cohort set
}
\description{
Collapse cohort records within a certain number of days
}
