% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{cohort_expression_query_builder_section}
\alias{cohort_expression_query_builder_section}
\title{Cohort Expression Query Builder}
\description{
Main SQL query builder for cohort expressions.
Generates complete cohort SQL from cohort definition.
}
\keyword{internal}
