% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_cohort_transformations.R
\name{cohort_first}
\alias{cohort_first}
\title{Keep only the earliest record for each person in a cohort}
\usage{
cohort_first(.data)
}
\arguments{
\item{.data}{A generated cohort set}
}
\value{
A lazy query on a generated cohort set
}
\description{
Keep only the earliest record for each person in a cohort
}
