% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortTransformations.R
\name{cohort_intersect}
\alias{cohort_intersect}
\title{Intersect all cohorts in a single cohort table}
\usage{
cohort_intersect(x, y)
}
\arguments{
\item{x}{A tbl reference to a cohort table with one or more cohorts}

\item{y}{A tbl reference to a cohort table with one cohort}
}
\value{
A lazy query that when executed will resolve to a new cohort table with
one cohort_definition_id resulting from the intersection of all cohorts x with the cohort in y
}
\description{
Intersect all cohorts in a single cohort table
}
