% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortTransformations.R
\name{cohort_last}
\alias{cohort_last}
\title{Keep only the latest record for each person in a cohort}
\usage{
cohort_last(x)
}
\arguments{
\item{x}{A generated cohort set}
}
\value{
A lazy query on a generated cohort set
}
\description{
Keep only the latest record for each person in a cohort
}
