% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortTransformations.R
\name{cohort_pad_end}
\alias{cohort_pad_end}
\title{Add or subtract days from the start or end of a cohort set}
\usage{
cohort_pad_end(.data, days = NULL, from = "end")
}
\arguments{
\item{.data}{A generated cohort set table reference}

\item{days}{The number of days to add. Can by any positive or negative integer}

\item{from}{Reference date to add or subtract days to. "start" or "end" (default)}
}
\value{
A lazy tbl query on a the cohort table
}
\description{
Add or subtract days from the start or end of a cohort set
}
