% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortTransformations.R
\name{cohort_union}
\alias{cohort_union}
\alias{cohortUnion}
\title{Union all cohorts in a cohort set with cohorts in a second cohort set}
\usage{
cohort_union(x, y)

cohortUnion(x, y)
}
\arguments{
\item{x}{A tbl reference to a cohort table with one or more generated cohorts}

\item{y}{A tbl reference to a cohort table with one generated cohort}
}
\value{
A lazy query that when executed will resolve to a new cohort table with
one the same cohort_definitions_ids in x resulting from the union of all cohorts
in x with the single cohort in y cohort table
}
\description{
Union all cohorts in a cohort set with cohorts in a second cohort set
}
