% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohortSet.R
\name{cohortAttrition}
\alias{cohortAttrition}
\title{Get attrition table from a GeneratedCohortSet object}
\usage{
cohortAttrition(x)
}
\arguments{
\item{x}{A generatedCohortSet object}
}
\description{
Get attrition table from a GeneratedCohortSet object
}
