% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{cohortExpressionFromJson}
\alias{cohortExpressionFromJson}
\title{Load cohort expression from JSON}
\usage{
cohortExpressionFromJson(json_str)
}
\arguments{
\item{json_str}{Character string containing cohort definition JSON (or path to JSON file)}
}
\value{
List structure representing the cohort expression
}
\description{
Load cohort expression from JSON
}
