% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohortSet.R
\name{cohortSet}
\alias{cohortSet}
\alias{cohort_set}
\title{Get cohort settings from a cohort_table object}
\usage{
cohortSet(x)

cohort_set(x)
}
\arguments{
\item{x}{A cohort_table object}
}
\description{
Get cohort settings from a cohort_table object

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
