% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortTransformations.R
\name{cohortUnion}
\alias{cohortUnion}
\alias{cohort_union}
\title{Union all cohorts in a cohort set with cohorts in a second cohort set}
\usage{
cohortUnion(x, y)

cohort_union(x, y)
}
\arguments{
\item{x}{A tbl reference to a cohort table with one or more generated cohorts}

\item{y}{A tbl reference to a cohort table with one generated cohort}
}
\value{
A lazy query that when executed will resolve to a new cohort table with
one the same cohort_definitions_ids in x resulting from the union of all cohorts
in x with the single cohort in y cohort table
}
\description{
Union all cohorts in a cohort set with cohorts in a second cohort set

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
