% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{collect.cdm_reference}
\alias{collect.cdm_reference}
\title{Bring a remote CDM reference into R}
\usage{
\method{collect}{cdm_reference}(x, ...)
}
\arguments{
\item{x}{A cdm_reference object.}

\item{...}{Not used. Included for compatibility.}
}
\value{
A cdm_reference object that is a list of R dataframes.
}
\description{
This function calls collect on a list of lazy queries and returns
the result as a list of dataframes.
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
vocab <- cdm_from_con(con, cdm_tables = c("concept", "concept_ancestor"))
local_vocab <- collect(vocab)
DBI::dbDisconnect(con, shutdown = TRUE)
}
}
