% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{collect.cdm_tbl}
\alias{collect.cdm_tbl}
\title{Bring a remote CDM table into R}
\usage{
\method{collect}{cdm_tbl}(x, ...)
}
\arguments{
\item{x}{A cdm_tbl object.}

\item{...}{Not used. Included for compatibility.}
}
\value{
A cdm_tbl object that is a R dataframe.
}
\description{
This function calls collect on a lazy query and returns
the result as a dataframe, ensuring that the column names are in lower case.
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
cdm <- cdm_from_con(con, "main")

local_concept <- collect(cdm$concept)
DBI::dbDisconnect(con, shutdown = TRUE)
}
}
