% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{collect_batch_used_domains_from_cohorts}
\alias{collect_batch_used_domains_from_cohorts}
\title{Collect CDM table names used by cohort objects (for ##*_FILTERED)}
\usage{
collect_batch_used_domains_from_cohorts(cohort_list)
}
\arguments{
\item{cohort_list}{List of cohort expression lists.}
}
\value{
Character vector of CDM table names (e.g. "DRUG_EXPOSURE").
}
\description{
Collect CDM table names used by cohort objects (for ##*_FILTERED)
}
\keyword{internal}
