% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohortSet.R
\name{computeAttrition}
\alias{computeAttrition}
\title{Compute the attrition for a set of cohorts}
\usage{
computeAttrition(cdm, cohortStem, cohortSet, cohortId = NULL)
}
\arguments{
\item{cdm}{A cdm reference created by CDMConnector.}

\item{cohortStem}{Stem for the cohort tables.}

\item{cohortSet}{Cohort set of the generated tables.}

\item{cohortId}{Cohort definition id of the cohorts that we want to generate
the attrition. If NULL all cohorts from cohort set will be used.}
}
\value{
the attrition as a data.frame
}
\description{
This function computes the attrition for a set of cohorts. It
uses the inclusion_result table so the cohort should be previously generated
using stats = TRUE.
}
