% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{computePermanent}
\alias{computePermanent}
\alias{compute_permanent}
\title{Run a dplyr query and store the result in a permanent table}
\usage{
computePermanent(x, name, schema = NULL, overwrite = FALSE)

compute_permanent(x, name, schema = NULL, overwrite = FALSE)
}
\arguments{
\item{x}{A dplyr query}

\item{name}{Name of the table to be created}

\item{schema}{Schema to create the new table in
Can be a length 1 or 2 vector.
(e.g. schema = "my_schema", schema = c("my_schema", "dbo"))}

\item{overwrite}{If the table already exists in the remote database
should it be overwritten? (TRUE or FALSE)}
}
\value{
A dplyr reference to the newly created table
}
\description{
This function has been superceded by \code{computeQuery} which should be used
instead of \code{computePermanent}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
