% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Caprification.R
\name{createConceptBasedCaprCohort}
\alias{createConceptBasedCaprCohort}
\title{Title}
\usage{
createConceptBasedCaprCohort(
  conceptSet,
  limit = "first",
  requiredObservation = c(0, 0),
  end = "observation_period_end_date",
  endArgs = list(conceptSet = conceptSet, persistenceWindow = 30L, surveillanceWindow =
    0L, daysSupplyOverride = NULL, index = c("startDate"), offsetDays = 7)
)
}
\arguments{
\item{endArgs}{}
}
\description{
Title
}
