% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{date_string_to_sql}
\alias{date_string_to_sql}
\title{Convert date string to SQL DATEFROMPARTS}
\usage{
date_string_to_sql(date_string)
}
\arguments{
\item{date_string}{Date in YYYY-MM-DD format}
}
\value{
SQL expression
}
\description{
Convert date string to SQL DATEFROMPARTS
}
