% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{dbms}
\alias{dbms}
\title{Get the database management system (dbms) from a cdm_reference or DBI connection}
\usage{
dbms(con)
}
\arguments{
\item{con}{A DBI connection or cdm_reference}
}
\value{
A character string representing the dbms that can be used with SqlRender
}
\description{
Get the database management system (dbms) from a cdm_reference or DBI connection
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
cdm <- cdm_from_con(con)
dbms(cdm)
dbms(con)
}
}
