% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eunomia.R
\name{download_optional_data}
\alias{download_optional_data}
\title{Download the example Eunomia CDM dataset}
\usage{
download_optional_data()
}
\description{
This function is deprecated. Use downloadEunomiaData instead.
}
