% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{dropStemTables}
\alias{dropStemTables}
\title{Drop tables starting with a prefix/stem}
\usage{
dropStemTables(cdm, stem, verbose = FALSE)
}
\arguments{
\item{cdm}{A CDM reference object}

\item{stem}{The prefix/stem identifying tables to drop. For example,
a stem of "study_results" would lead to dropping any table starting with this
(ie both "study_results" and "study_results_analysis_1" would be dropped if
they exist in the write schema). Prefix will be interpreted as a regular
expression.}

\item{verbose}{Print a message when dropping a table? TRUE or FALSE (default)}
}
\value{
Invisibly returns the cdm object
}
\description{
Drop tables starting with a prefix/stem
}
