% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSource.R
\name{dropTable.db_cdm}
\alias{dropTable.db_cdm}
\title{Drop table from a database backed cdm object}
\usage{
dropTable.db_cdm(cdm, name)
}
\arguments{
\item{cdm}{a cdm_reference object}

\item{name}{A character vector of table names to be dropped}
}
\description{
Tables will be dropped from the write schema of the cdm.
}
