% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{eunomia_dir}
\alias{eunomia_dir}
\title{Create a new Eunomia CDM}
\usage{
eunomia_dir(exdir = tempdir())
}
\arguments{
\item{exdir}{Enclosing directory where the Eunomia CDM should be created}
}
\value{
The full path to the new Eunomia CDM that can be passed to \code{dbConnect()}
}
\description{
Create a copy of the duckdb Eunomia CDM and return the file path
}
\examples{
\dontrun{
library(DBI)
library(CDMConnector)
con <- dbConnect(duckdb::duckdb(), dbdir = getEunomiaPath())
dbListTables(con)
dbDisconnect(con)
}
}
