% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_data.R
\name{eunomia_is_available}
\alias{eunomia_is_available}
\title{Has the eunomia dataset been cached?}
\usage{
eunomia_is_available()
}
\value{
TRUE if the eunomia example dataset is available and FASLE otherwise
}
\description{
Has the eunomia dataset been cached?
}
