% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateadd.R
\name{extract}
\alias{extract}
\title{Extract the day, month or year of a date in sql.}
\usage{
extract(date, value = "year", dms = NULL)
}
\arguments{
\item{date}{Character variable that points to a date column.}

\item{value}{Value to extract. It can be "year", "month", or "day".}

\item{dms}{Database manager system, if NULL it is auto detected.}
}
\description{
Extract the day, month or year of a date in sql.
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb(), ":memory:")
date_tbl <- dplyr::copy_to(con,
                           data.frame(birth_date = as.Date("1993-04-19")),
                           name = "tmp",
                           temporary = TRUE)
df <- date_tbl \%>\%
  dplyr::mutate(year = !!extract("birth_date", "year")) \%>\%
  dplyr::mutate(month = !!extract("birth_date", "month")) \%>\%
  dplyr::mutate(day = !!extract("birth_date", "day")) \%>\%
  dplyr::collect()
DBI::dbDisconnect(con, shutdown = TRUE)
}
}
