% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{extract_criteria}
\alias{extract_criteria}
\title{Extract criteria type and data from polymorphic criteria structure}
\usage{
extract_criteria(item)
}
\arguments{
\item{item}{List - either named by type (ConditionOccurrence=...) or in Criteria/criteria}
}
\value{
List with type, data
}
\description{
Extract criteria type and data from polymorphic criteria structure
}
