% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate,cdm_reference,CohortSet-method}
\alias{generate,cdm_reference,CohortSet-method}
\title{Generate a set of cohorts}
\usage{
\S4method{generate}{cdm_reference,CohortSet}(cdm, x, cohortTableName = "cohort", overwrite = FALSE)
}
\arguments{
\item{cdm}{cdm reference object}

\item{x}{A cohort definition set dataframe}

\item{cohortTableName}{The name of the cohort table in the cdm. Defaults to 'cohort'.}
}
\value{
cdm reference object with the added cohort table containing generated cohorts
}
\description{
This function generates a set of cohorts in the cohort table.
}
