% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate a specification on a CDM}
\usage{
generate(cdm, x, ...)
}
\arguments{
\item{cdm}{A cdm reference}

\item{x}{A specification of an analysis or cohort set}

\item{...}{Additional parameters passed to methods}
}
\value{
A cdm reference with the results tables added
}
\description{
\code{generate} is a generic function that accepts a cdm and a specification of an
analysis or cohort set to be generated.
}
