% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generateCohortSet}
\alias{generateCohortSet}
\title{Generate a set of cohorts}
\usage{
generateCohortSet(cdm, cohortDefinitionSet, cohortTableName)
}
\arguments{
\item{cdm}{cdm reference object}

\item{CohortDefinitionSet}{A cohort definition set dataframe}
}
\value{
cdm reference object with the added cohort table containing generated cohorts
}
\description{
This function generates a set of cohorts in the cohort table.
}
