% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateConceptCohortSet.R
\name{generateConceptCohortSet}
\alias{generateConceptCohortSet}
\alias{generate_concept_cohort_set}
\title{Create a new generated cohort set from a list of concept sets}
\usage{
generateConceptCohortSet(
  cdm,
  conceptSet = NULL,
  name,
  limit = "first",
  requiredObservation = c(0, 0),
  end = "observation_period_end_date",
  subsetCohort = NULL,
  subsetCohortId = NULL,
  overwrite = TRUE,
  endArgs = list(persistenceWindow = 30L, surveillanceWindow = 0L, daysSupplyOverride =
    NULL, index = c("startDate"), offsetDays = dplyr::if_else(is.numeric(end), end, 7)),
  containsSourceConceptIds = FALSE
)

generate_concept_cohort_set(
  cdm,
  concept_set = NULL,
  name = "cohort",
  limit = "first",
  required_observation = c(0, 0),
  end = "observation_period_end_date",
  subset_cohort = NULL,
  subset_cohort_id = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{cdm}{A cdm reference object created by \code{CDMConnector::cdmFromCon} or \code{CDMConnector::cdm_from_con}}

\item{conceptSet}{A named list of numeric vectors, Capr ConceptSet or a Concept Set Expression created \code{omopgenerics::newConceptSetExpression}}

\item{name}{The name of the new generated cohort table as a character string}

\item{limit}{Include "first" (default) or "all" occurrences of events in the cohort
\itemize{
\item{"first" will include only the first occurrence of any event in the concept set in the cohort.}
\item{"all" will include all occurrences of the events defined by the concept set in the cohort.}
\item{"latest" will include latest occurrence of the events defined by the concept set in the cohort.}
}}

\item{requiredObservation}{A numeric vector of length 2 that specifies the number of days of required}

\item{end}{How should the \code{cohort_end_date} be defined?
\itemize{
\item{"observation_period_end_date" (default): The earliest observation_period_end_date after the event start date}
\item{"drug_exit"}: End of continuous drug expose
\item{"event_end_date"}: The event end date is derived from adding a number of days to the event's start or end date.
}}

\item{overwrite}{Should the cohort table be overwritten if it already exists? TRUE (default) or FALSE.}

\item{endArgs}{Set \code{cohort_end_date} strategy arguments
\itemize{
\item{"persistenceWindow" (30 by default)}: allow for a maximum of days between exposure records when inferring the era of persistence exposure
\item{"surveillanceWindow" (0 by default)}: add days to the end of the era of persistence exposure as an additional period of surveillance prior to cohort exit.
\item{"daysSupplyOverride" (NULL by default)}: Use days supply and exposure end date for exposure duration.
\item{"index" ("startDate" by default)}: Event date to offset from
\item{"offsetDays" (captures end by default if numeric or 7 by default)}: Number of days offset
}}

\item{containsSourceConceptIds}{Boolean value. Point if concept set contains source codes to search}
}
\value{
A cdm reference object with the new generated cohort set table added
}
\description{
Generate a new cohort set from one or more concept sets. Each
concept set will result in one cohort and represent the time during which
the concept was observed for each subject/person. Concept sets can be
passed to this function as:
\itemize{
\item{A named list of numeric vectors, one vector per concept set}
\item{A named list of Capr concept sets}
}

Clinical observation records will be looked up in the respective domain tables
using the vocabulary in the CDM. If a required domain table does not exist in
the cdm object a warning will be given.
Concepts that are not in the vocabulary or in the data will be silently ignored.
If end dates are missing or do not exist, as in the case of the procedure and
observation domains, the the start date will be used as the end date.
}
