% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_codeset_clause}
\alias{get_codeset_clause}
\title{Get codeset join for criteria with codeset_id}
\usage{
get_codeset_clause(criteria, standard_col, source_col = NULL)
}
\arguments{
\item{criteria}{List}

\item{standard_col}{Standard concept column}

\item{source_col}{Source concept column (optional)}
}
\value{
SQL fragment
}
\description{
Get codeset join for criteria with codeset_id
}
