% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_codeset_in_expression}
\alias{get_codeset_in_expression}
\title{Get codeset IN expression}
\usage{
get_codeset_in_expression(codeset_id, column_name, is_exclusion = FALSE)
}
\arguments{
\item{codeset_id}{Integer}

\item{column_name}{Column name}

\item{is_exclusion}{Logical}
}
\value{
SQL expression
}
\description{
Get codeset IN expression
}
