% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_codeset_join_expression}
\alias{get_codeset_join_expression}
\title{Get codeset join expression}
\usage{
get_codeset_join_expression(
  standard_codeset_id,
  standard_concept_column,
  source_codeset_id,
  source_concept_column
)
}
\arguments{
\item{standard_codeset_id}{Integer or NULL}

\item{standard_concept_column}{Column name}

\item{source_codeset_id}{Integer or NULL}

\item{source_concept_column}{Column name}
}
\value{
SQL JOIN clause
}
\description{
Get codeset join expression
}
