% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_concept_ids}
\alias{get_concept_ids}
\title{Get concept IDs from concept list}
\usage{
get_concept_ids(concepts)
}
\arguments{
\item{concepts}{List of concept objects}
}
\value{
Integer vector of concept IDs
}
\description{
Get concept IDs from concept list
}
