% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_concept_sets}
\alias{get_concept_sets}
\title{Get concept sets from cohort expression}
\usage{
get_concept_sets(cohort)
}
\arguments{
\item{cohort}{List (cohort expression)}
}
\value{
List of concept sets
}
\description{
Get concept sets from cohort expression
}
\keyword{internal}
