% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_corelated_criteria_query}
\alias{get_corelated_criteria_query}
\title{Get correlated criteria query for a criteria list item (Criteria + StartWindow + Occurrence)}
\usage{
get_corelated_criteria_query(
  correlated_item,
  event_table,
  index_id,
  cdm_schema
)
}
\description{
Get correlated criteria query for a criteria list item (Criteria + StartWindow + Occurrence)
}
