% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_criteria_list}
\alias{get_criteria_list}
\title{Get criteria list from primary criteria}
\usage{
get_criteria_list(pc)
}
\arguments{
\item{pc}{List (primary criteria)}
}
\value{
List of criteria items
}
\description{
Get criteria list from primary criteria
}
\keyword{internal}
