% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_criteria_sql}
\alias{get_criteria_sql}
\title{Dispatch to appropriate criteria SQL builder}
\usage{
get_criteria_sql(criteria, cdm_schema = "@cdm_database_schema")
}
\arguments{
\item{criteria}{List - criteria object (may be wrapped in type key)}

\item{cdm_schema}{CDM schema name}
}
\value{
SQL string
}
\description{
Dispatch to appropriate criteria SQL builder
}
