% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_date_adjustment_expression}
\alias{get_date_adjustment_expression}
\title{Get date adjustment expression}
\usage{
get_date_adjustment_expression(date_adjustment, start_column, end_column)
}
\arguments{
\item{date_adjustment}{List with startOffset, endOffset, startWith, endWith}

\item{start_column}{Column for start}

\item{end_column}{Column for end}
}
\value{
SQL expression
}
\description{
Get date adjustment expression
}
