% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_demographic_criteria_query}
\alias{get_demographic_criteria_query}
\title{Get demographic criteria query}
\usage{
get_demographic_criteria_query(
  dc,
  event_table,
  index_id,
  cdm_schema = "@cdm_database_schema"
)
}
\arguments{
\item{dc}{List - DemographicCriteria (OccurrenceStartDate, OccurrenceEndDate, Age, Gender, etc.)}

\item{event_table}{Event table expression}

\item{index_id}{Index for this criteria}

\item{cdm_schema}{CDM schema}
}
\description{
Get demographic criteria query
}
