% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_domain_concept_col}
\alias{get_domain_concept_col}
\title{Get domain concept column for criteria type (for COUNT DISTINCT)}
\usage{
get_domain_concept_col(criteria)
}
\arguments{
\item{criteria}{Criteria item (list).}
}
\description{
Get domain concept column for criteria type (for COUNT DISTINCT)
}
