% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlrender.R
\name{get_key}
\alias{get_key}
\title{Get value from list with fallback keys (PascalCase/camelCase)}
\usage{
get_key(obj, keys, default = NULL)
}
\arguments{
\item{obj}{List}

\item{keys}{Character vector of possible key names}

\item{default}{Default value if not found}
}
\value{
Value or default
}
\description{
Get value from list with fallback keys (PascalCase/camelCase)
}
\keyword{internal}
