% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_observation_window}
\alias{get_observation_window}
\title{Get observation window from primary criteria}
\usage{
get_observation_window(pc)
}
\arguments{
\item{pc}{List (primary criteria)}
}
\value{
List with prior_days, post_days
}
\description{
Get observation window from primary criteria
}
\keyword{internal}
