% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_operator}
\alias{get_operator}
\title{Get SQL operator from op string}
\usage{
get_operator(op)
}
\arguments{
\item{op}{Operator name (lt, lte, eq, gt, gte, etc.)}
}
\value{
SQL operator string
}
\description{
Get SQL operator from op string
}
