% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_primary_criteria}
\alias{get_primary_criteria}
\title{Get primary criteria from cohort expression}
\usage{
get_primary_criteria(cohort)
}
\arguments{
\item{cohort}{List (cohort expression)}
}
\value{
List or NULL
}
\description{
Get primary criteria from cohort expression
}
\keyword{internal}
