% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_primary_limit_type}
\alias{get_primary_limit_type}
\title{Get primary limit type from primary criteria}
\usage{
get_primary_limit_type(pc)
}
\arguments{
\item{pc}{List (primary criteria)}
}
\value{
Character "All", "First", or "Last"
}
\description{
Get primary limit type from primary criteria
}
\keyword{internal}
