% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFullTableNameQuoted}
\alias{getFullTableNameQuoted}
\title{Get the full table name consisting of the schema and table name.}
\usage{
getFullTableNameQuoted(x, name, schema)
}
\arguments{
\item{x}{A dplyr query}

\item{name}{Name of the table to be created.}

\item{schema}{Schema to create the new table in
Can be a length 1 or 2 vector.
(e.g. schema = "my_schema", schema = c("my_schema", "dbo"))}
}
\value{
the full table name
}
\description{
Get the full table name consisting of the schema and table name.
}
\examples{
\dontrun{
library(CDMConnector)
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
cdm <- cdm_from_con(con)
schema <- c("my_schema", "dbo")
name <- "myTable"
getFullTableNameQuoted(cdm$person, name, schema)
}
}
