% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlrender.R
\name{or-or}
\alias{or-or}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left value}

\item{y}{Right value (used when x is NULL)}
}
\value{
x or y
}
\description{
Null coalescing operator
}
\keyword{internal}
