% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{print.cdm_reference}
\alias{print.cdm_reference}
\title{Print a CDM reference object}
\usage{
\method{print}{cdm_reference}(x, ...)
}
\arguments{
\item{x}{A cdm_reference object}

\item{...}{Included for compatiblity with generic. Not used.}
}
\value{
Invisibly returns the input
}
\description{
Print a CDM reference object
}
