% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCohortSet.R
\name{readCohortSet}
\alias{readCohortSet}
\title{Read a set of cohort definitions into R}
\usage{
readCohortSet(path)
}
\arguments{
\item{path}{The path to a folder containing Circe cohort definition json files
and optionally a csv file named CohortsToCreate.csv with columns cohortId, cohortName, and jsonPath.}
}
\description{
A "cohort set" is a collection of cohort definitions. In R this is stored in a dataframe with cohortId, cohortName, cohort, and sql columns.
On disk this is stored as a folder with a CohortsToCreate.csv file and one or more json files.
If the CohortsToCreate.csv file is missing then all of the json files in the folder will be used,
cohortIds will be automatically assigned in alphabetical order, and cohortNames will match the file names.
}
