% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{remote_con}
\alias{remote_con}
\title{Get the database connection from a cdm object or remote table reference}
\usage{
remote_con(x)
}
\arguments{
\item{x}{A cdm_reference or tbl_sql object}
}
\value{
A DBI connection
}
\description{
Get the database connection from a cdm object or remote table reference
}
