% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eunomia.R
\name{requireEunomia}
\alias{requireEunomia}
\title{Require eunomia to be available. The function makes sure that you can later
create a eunomia database with \code{eunomiaDir()}.}
\usage{
requireEunomia(datasetName = "GiBleed", cdmVersion = "5.3")
}
\arguments{
\item{datasetName}{Name of the Eunomia dataset to check. Defaults to "GiBleed".}

\item{cdmVersion}{Version of the Eunomia dataset to check. Must be "5.3" or "5.4".}
}
\value{
Path to eunomia database.
}
\description{
Require eunomia to be available. The function makes sure that you can later
create a eunomia database with \code{eunomiaDir()}.
}
