% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{split_in_clause}
\alias{split_in_clause}
\title{Split IN clause for large value lists (Oracle 1000 limit)}
\usage{
split_in_clause(column_name, values, max_length = 1000)
}
\arguments{
\item{column_name}{SQL column name}

\item{values}{Integer vector}

\item{max_length}{Max values per chunk}
}
\value{
SQL expression string
}
\description{
Split IN clause for large value lists (Oracle 1000 limit)
}
