% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{[.cdm_reference}
\alias{[.cdm_reference}
\title{Subset a cdm reference object}
\usage{
\method{[}{cdm_reference}(x, i)
}
\arguments{
\item{x}{A cdm reference}

\item{i}{The name or index of the table to extract from the cdm object}
}
\value{
A single cdm table reference
}
\description{
Subset a cdm reference object
}
