% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlrender.R
\name{translate_r}
\alias{translate_r}
\title{Translate SQL to target dialect
(Pure R implementation — kept as fallback)}
\usage{
translate_r(
  sql,
  targetDialect,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  oracleTempSchema = NULL
)
}
\description{
Translate SQL to target dialect
(Pure R implementation — kept as fallback)
}
