% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-deprecated.R
\name{union_cohorts}
\alias{union_cohorts}
\alias{unionCohorts}
\title{Union all cohorts in a single cohort table}
\usage{
union_cohorts(x, cohort_definition_id = 1L)

unionCohorts(x, cohort_definition_id = 1L)
}
\arguments{
\item{x}{A tbl reference to a cohort table}

\item{cohort_definition_id}{A number to use for the new cohort_definition_id

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}}
}
\value{
A lazy query that when executed will resolve to a new cohort table with
one cohort_definition_id resulting from the union of all cohorts in the original
cohort table
}
\description{
Union all cohorts in a single cohort table
}
