% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{unique_table_name}
\alias{unique_table_name}
\title{Create a unique table name for temp tables}
\usage{
unique_table_name()
}
\value{
A string that can be used as a dbplyr temp table name
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
