% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{uniqueTableName}
\alias{uniqueTableName}
\alias{unique_table_name}
\title{Create a unique table name for temp tables}
\usage{
uniqueTableName()

unique_table_name()
}
\value{
A string that can be used as a dbplyr temp table name
}
\description{
Create a unique table name for temp tables
}
