% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{version}
\alias{version}
\title{Get the CDM version}
\usage{
version(cdm)
}
\arguments{
\item{cdm}{A cdm object}
}
\value{
"5.3", "5.4" or "5.5"
}
\description{
Extract the CDM version attribute from a cdm_reference object
}
\examples{
\dontrun{
library(CDMConnector)
con <- DBI::dbConnect(duckdb::duckdb(), eunomiaDir())
cdm <- cdmFromCon(con, cdmSchema = "main", writeSchema = "main")
version(cdm)

DBI::dbDisconnect(con, shutdown = TRUE)
}
}
