% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{wrap_criteria_query}
\alias{wrap_criteria_query}
\title{Wrap base criteria query with correlated criteria group}
\usage{
wrap_criteria_query(base_query, group, cdm_schema = "@cdm_database_schema")
}
\description{
Wrap base criteria query with correlated criteria group
}
