% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkCDMConnector.R
\name{benchmarkCDMConnector}
\alias{benchmarkCDMConnector}
\title{Run benchmark of tasks using CDMConnector}
\usage{
benchmarkCDMConnector(cdm)
}
\arguments{
\item{cdm}{A CDM reference object}
}
\value{
a tibble with time taken for different analyses
}
\description{
Run benchmark of tasks using CDMConnector
}
\examples{
\dontrun{
library(CDMConnector)
con <- DBI::dbConnect(duckdb::duckdb(), eunomia_dir())
cdm <- cdm_from_con(con, cdm_schema = "main", write_schema = "main")
benchmarkCDMConnector(cdm)

DBI::dbDisconnect(con, shutdown = TRUE)
}
}
