% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_device_exposure_sql}
\alias{build_device_exposure_sql}
\title{Build Device Exposure SQL}
\usage{
build_device_exposure_sql(criteria, cdm_schema = "@cdm_database_schema")
}
\description{
Build Device Exposure SQL
}
