% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_drug_exposure_sql}
\alias{build_drug_exposure_sql}
\title{Build Drug Exposure SQL}
\usage{
build_drug_exposure_sql(criteria, cdm_schema = "@cdm_database_schema")
}
\arguments{
\item{criteria}{Criteria list.}

\item{cdm_schema}{CDM schema name.}
}
\description{
Build Drug Exposure SQL
}
