% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_location_region_sql}
\alias{build_location_region_sql}
\title{Build Location Region SQL}
\usage{
build_location_region_sql(criteria, cdm_schema = "@cdm_database_schema")
}
\arguments{
\item{criteria}{Criteria list.}

\item{cdm_schema}{CDM schema name.}
}
\description{
Build Location Region SQL
}
