% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{build_window_criteria}
\alias{build_window_criteria}
\title{Build window criteria clause from StartWindow/EndWindow}
\usage{
build_window_criteria(start_window, end_window, check_observation_period)
}
\description{
Build window criteria clause from StartWindow/EndWindow
}
