% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{cdm_from_con}
\alias{cdm_from_con}
\title{Create a CDM reference reference to a database connection}
\usage{
cdm_from_con(
  con,
  cdm_schema = NULL,
  select = tbl_group("default"),
  write_schema = NULL
)
}
\arguments{
\item{con}{A DBI database connection to a database where an OMOP CDM v5.4 instance is located.}

\item{cdm_schema}{The schema where the OMOP CDM tables are located. Defaults to NULL.}

\item{select}{Which tables should be included?
\itemize{
\item{"all"}{all CDM tables}
\item{"vocab"}{the CDM vocabulary tables}
\item{"clinical"}{the clinical CDM tables}
}}

\item{write_schema}{An optional schema in the CDM database that the user has write access to.}
}
\value{
A list of dplyr database table references pointing to CDM tables
}
\description{
Create a CDM reference reference to a database connection
}
