% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{cdmCon}
\alias{cdmCon}
\title{Get underlying database connection}
\usage{
cdmCon(cdm)
}
\arguments{
\item{cdm}{A cdm reference object created by \code{cdm_from_con}}
}
\value{
A reference to the database containing tables in the cdm reference
}
\description{
Get underlying database connection
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())

cdm <- cdm_from_con(con = con, cdm_name = "Eunomia",
                    cdm_schema =  "main", write_schema = "main")

cdmCon(cdm)

DBI::dbDisconnect(con)
}
}
