% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMConnector-package.R
\docType{package}
\name{CDMConnector-package}
\alias{CDMConnector}
\alias{CDMConnector-package}
\title{CDMConnector: Connect to an OMOP Common Data Model}
\description{
Provides tools for working with observational health data in the 
  Observational Medical Outcomes Partnership (OMOP) Common Data Model format with a pipe friendly syntax.
  Common data model database table references are stored in a single compound object along with metadata.
}
\author{
\strong{Maintainer}: Adam Black \email{black@ohdsi.org} (\href{https://orcid.org/0000-0001-5576-8701}{ORCID})

}
\keyword{internal}
