% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSource.R
\name{cdmDisconnect.db_cdm}
\alias{cdmDisconnect.db_cdm}
\title{Disconnect the connection of the cdm object}
\usage{
\method{cdmDisconnect}{db_cdm}(cdm, dropWriteSchema = FALSE, ...)
}
\arguments{
\item{cdm}{cdm reference}

\item{dropWriteSchema}{Whether to drop tables in the writeSchema}
}
\description{
This function will disconnect from the database as well as drop
"temporary" tables that were created on database systems that do not support
actual temporary tables. Currently temp tables are emulated on
Spark/Databricks systems.
}
