% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{cdmDisconnect}
\alias{cdmDisconnect}
\alias{cdm_disconnect}
\title{Disconnect the connection of the cdm object}
\usage{
cdmDisconnect(cdm)

cdm_disconnect(cdm)
}
\arguments{
\item{cdm}{cdm reference}
}
\description{
Disconnect the connection of the cdm object
}
