% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{cdmWriteSchema}
\alias{cdmWriteSchema}
\title{Get cdm write schema}
\usage{
cdmWriteSchema(cdm)
}
\arguments{
\item{cdm}{A cdm reference object created by \code{cdmFromCon}}
}
\value{
The database write schema
}
\description{
Get cdm write schema
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomiaDir())

cdm <- cdmFromCon(con = con, cdmName = "Eunomia",
                    cdmSchema =  "main", writeSchema = "main")

cdmWriteSchema(cdm)

DBI::dbDisconnect(con)
}
}
