% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohortSet.R
\name{cohort_count}
\alias{cohort_count}
\title{Get cohort counts from a generated_cohort_set object.}
\usage{
cohort_count(cohort)
}
\arguments{
\item{cohort}{A generated_cohort_set object.}
}
\value{
A table with the counts.
}
\description{
Get cohort counts from a generated_cohort_set object.
}
