% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohortSet.R
\name{cohort_count}
\alias{cohort_count}
\title{Get cohort counts from a generated_cohort_set object.}
\usage{
cohort_count(cohort)
}
\arguments{
\item{cohort}{A generated_cohort_set object.}
}
\value{
A table with the counts.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\dontrun{
library(CDMConnector)
library(dplyr)

con <- DBI::dbConnect(duckdb::duckdb(), eunomiaDir())
cdm <- cdmFromCon(con = con, cdmSchema = "main", writeSchema = "main")
cdm <- generateConceptCohortSet(
  cdm = cdm, conceptSet = list(pharyngitis = 4112343), name = "new_cohort"
)
cohortCount(cdm$new_cohort)
}
}
