% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_cohort_transformations.R
\name{cohort_intersect}
\alias{cohort_intersect}
\title{Intersect all cohorts in a single cohort table}
\usage{
cohort_intersect(x, y)
}
\arguments{
\item{.data}{A tbl reference to a cohort table}

\item{cohort_definition_id}{A positive integer to use for the new cohort_definition_id}
}
\value{
A lazy query that when executed will resolve to a new cohort table with
one cohort_definition_id resulting from the intersection of all cohorts in the original
cohort table
}
\description{
Intersect all cohorts in a single cohort table
}
