% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohortSet.R
\name{cohortAttrition}
\alias{cohortAttrition}
\title{Return attrition table from a generated cohort set object}
\usage{
cohortAttrition(x)
}
\arguments{
\item{x}{A generatedCohortSet object}
}
\description{
Return attrition table from a generated cohort set object
}
