% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohortSet.R
\name{cohortAttrition}
\alias{cohortAttrition}
\alias{cohort_attrition}
\title{Get attrition table from a cohort_table object}
\usage{
cohortAttrition(x)

cohort_attrition(x)
}
\arguments{
\item{x}{A cohort_table object}
}
\description{
Get attrition table from a cohort_table object

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
