% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohortSet.R
\name{cohortSet}
\alias{cohortSet}
\title{Get cohort settings from a GeneratedCohortSet object}
\usage{
cohortSet(x)
}
\arguments{
\item{x}{A generatedCohortSet object}
}
\description{
Get cohort settings from a GeneratedCohortSet object
}
