% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{dropTable}
\alias{dropTable}
\title{Drop tables from write_schema of a cdm object}
\usage{
dropTable(cdm, name)
}
\arguments{
\item{cdm}{A cdm reference}

\item{name}{A character vector of tables in the cdm's write_schema}
}
\value{
A modified cdm reference with the tables removed
}
\description{
cdm objects can have zero or more cohort tables stored in a special schema
where the user has write access. This function removes tables from a cdm's
write_schema
}
\examples{
\dontrun{
library(CDMConnector)

con <- DBI::dbConnect(duckdb::duckdb(), dbdir = CDMConnector::eunomia_dir())
cdm <- cdm_from_con(con, "main")

# create a temporary table in the remote database from a query
cdm$tmp_table <- cdm$concept \%>\%
  dplyr::count(domain_id == "Drug") \%>\%
  computeQuery("tmp_table", temporary = FALSE, schema = "main")

"tmp_table" \%in\% DBI::dbListTables(con)
# [1] TRUE
"tmp_table" \%in\% names(cdm)
# [1] TRUE

cdm <- dropTable(cdm, "tmp_table")

"tmp_table" \%in\% DBI::dbListTables(con)
# [1] FALSE
"tmp_table" \%in\% names(cdm)
# [1] FALSE

DBI::dbDisconnect(con, shutdown = TRUE)
}
}
