% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_data.R
\name{eunomia_dir}
\alias{eunomia_dir}
\alias{eunomiaDir}
\title{Create a new Eunomia CDM}
\usage{
eunomia_dir(exdir = NULL)

eunomiaDir(exdir = NULL)
}
\arguments{
\item{exdir}{Enclosing directory where the Eunomia CDM should be created.
If NULL (default) then a temp folder is created.}
}
\value{
The full path to the new Eunomia CDM that can be passed to \code{dbConnect()}
}
\description{
Create a copy of the duckdb Eunomia CDM and return the file path
}
\section{Functions}{
\itemize{
\item \code{eunomiaDir()}: camelCase alias

}}
\examples{
\dontrun{
library(DBI)
library(CDMConnector)
con <- dbConnect(duckdb::duckdb(), dbdir = getEunomiaPath())
dbListTables(con)
dbDisconnect(con)
}
}
