% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eunomia.R
\name{eunomia_is_available}
\alias{eunomia_is_available}
\alias{eunomiaIsAvailable}
\title{Has the Eunomia dataset been cached?}
\usage{
eunomia_is_available(dataset_name = "GiBleed", cdm_version = "5.3")

eunomiaIsAvailable(datasetName = "GiBleed", cdmVersion = "5.3")
}
\arguments{
\item{dataset_name, datasetName}{Name of the Eunomia dataset to check. Defaults to "GiBleed".}

\item{cdm_version, cdmVersion}{Version of the Eunomia dataset to check. Must be "5.3" or "5.4".}
}
\value{
TRUE if the eunomia example dataset is available and FASLE otherwise
}
\description{
Has the Eunomia dataset been cached?
}
