% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eunomia.R
\name{exampleDatasets}
\alias{exampleDatasets}
\title{List the available example CDM datasets}
\usage{
exampleDatasets()
}
\value{
A character vector with example CDM dataset identifiers
}
\description{
List the available example CDM datasets
}
\examples{
\dontrun{
library(CDMConnector)
exampleDatasets()[1]
#> [1] "GiBleed"

con <- DBI::dbConnect(duckdb::duckdb(), eunomiaDir("GiBleed"))
cdm <- cdmFromCon(con)
}
}
