% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{generateCohortSet2}
\alias{generateCohortSet2}
\title{Generate Cohort set 2}
\usage{
generateCohortSet2(cdm, cohortSet, name, cache = FALSE)
}
\arguments{
\item{cdm}{cdm}

\item{cohortSet}{cohortSet}

\item{name}{cohort table name}

\item{cache}{Logical; if TRUE, enable incremental DAG caching. Previously
computed intermediate tables are reused, and only changed portions of the
DAG are recomputed. Default FALSE.}
}
\value{
a cdm with the cohort table
}
\description{
Generate Cohort set 2
}
