% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateConceptCohortSet.R
\name{generateConceptCohortSet}
\alias{generateConceptCohortSet}
\alias{generate_concept_cohort_set}
\title{Internal function to create a new generated cohort set from a list of concept sets}
\usage{
generateConceptCohortSet(
  cdm,
  conceptSet = NULL,
  name = "cohort",
  limit = "first",
  end = "observation_period_end_date",
  overwrite = FALSE
)

generate_concept_cohort_set(
  cdm,
  concept_set = NULL,
  name = "cohort",
  end = "observation_period_end_date",
  overwrite = FALSE
)
}
\arguments{
\item{cdm}{A cdm reference object created by \code{CDMConnector::cdmFromCon} or \code{CDMConnector::cdm_from_con}}

\item{conceptSet, concept_set}{A named list of numeric vectors or Capr concept sets}

\item{name}{The name of the new generated cohort table as a character string}

\item{limit}{"first" (default) or "all".
\itemize{
\item{"first" will include only the first occurrence of any event in the concept set in the cohort.}
\item{"all" will include all occurrences of the events defined by the concept set in the cohort.}
}}

\item{end}{How should the \code{cohort_end_date} be defined?
\itemize{
\item{"observation_period_end_date" (default): The earliest observation_period_end_date after the event start date}
\item{numeric scalar: A fixed number of days from the event start date}
\item{"event_end_date"}: The event end date. If the event end date is not populated then the event start date will be used
}}

\item{overwrite}{Should the cohort table be overwritten if it already exists? TRUE or FALSE}
}
\value{
A cdm reference object with the new generated cohort set table added
}
\description{
Generate a new cohort set from one or more concept sets. Each
concept set will result in one cohort and represent the time during which
the concept was observed for each subject/person. Concept sets can be
passed to this function as:
\itemize{
\item{A named list of numeric vectors, one vector per concept set}
\item{A named list of Capr concept sets}
}

Clinical observation records will be looked up in the respective domain tables
using the vocabulary in the CDM. If a required domain table does not exist in
the cdm object a warning will be given.
Concepts that are not in the vocabulary or in the data will be silently ignored.
If end dates are missing or do not exist, as in the case of the procedure and
observation domains, the the start date will be used as the end date.
}
