% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateConceptCohortSet.R
\name{generateConceptCohortSet}
\alias{generateConceptCohortSet}
\title{Internal function to create a new generated cohort set from a list of concept sets}
\usage{
generateConceptCohortSet(
  cdm,
  conceptSet = NULL,
  name = "cohort",
  overwrite = FALSE
)
}
\arguments{
\item{cdm}{A cdm reference object created by \code{CDMConnector::cdmFromCon} or \code{CDMConnector::cdm_from_con}}

\item{conceptSet}{A named list of numeric vectors or Capr concept sets}

\item{name}{The name of the new generated cohort table as a character string}

\item{overwrite}{Should the cohort table be overwritten if it already exists? TRUE or FALSE}
}
\value{
A cdm reference object with the new generated cohort set table added
}
\description{
Generate a new cohort set from one or more concept sets. Each
concept set will result in one cohort and represent the time during which
the concept was observed for each subject/person. Concept sets can be
passed to this function as:
\itemize{
\item{A named list of numeric vectors, one vector per concept set}
\item{A named list of Capr concept sets}
}

Clinical observation records will be looked up in the respective domain tables
using the vocabulary in the CDM. If a required domain table does not exist in
the cdm object a warning will be given.
Concepts that are not in the vocabulary or in the data will be silently ignored.
If end dates are missing or do not exist, as in the case of the procedure and
observation domains, the the start date will be used as the end date.
}
