% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_corelated_criteria_query}
\alias{get_corelated_criteria_query}
\title{Get correlated criteria query for a criteria list item (Criteria + StartWindow + Occurrence)}
\usage{
get_corelated_criteria_query(
  correlated_item,
  event_table,
  index_id,
  cdm_schema
)
}
\arguments{
\item{correlated_item}{Correlated criteria item.}

\item{event_table}{Event table expression.}

\item{index_id}{Index id.}

\item{cdm_schema}{CDM schema name.}
}
\description{
Get correlated criteria query for a criteria list item (Criteria + StartWindow + Occurrence)
}
