% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_criteria_group_query}
\alias{get_criteria_group_query}
\title{Get criteria group query}
\usage{
get_criteria_group_query(
  group,
  event_table,
  cdm_schema = "@cdm_database_schema"
)
}
\arguments{
\item{group}{Criteria group (Type, CriteriaList, Groups, etc.).}

\item{event_table}{Event table expression.}

\item{cdm_schema}{CDM schema name.}
}
\description{
Get criteria group query
}
