% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe.R
\name{get_occurrence_operator}
\alias{get_occurrence_operator}
\title{Get occurrence operator from type (0=Exactly, 1=AtMost, 2=AtLeast)}
\usage{
get_occurrence_operator(occ_type)
}
\arguments{
\item{occ_type}{Occurrence type (0, 1, or 2).}
}
\description{
Get occurrence operator from type (0=Exactly, 1=AtMost, 2=AtLeast)
}
