% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inSchema}
\alias{inSchema}
\title{Helper for working with compound schema}
\usage{
inSchema(schema, table, dbms = NULL)
}
\arguments{
\item{schema}{A schema name as a character string}

\item{table}{A table name as character string}

\item{dbms}{The name of the database management system as returned
by \code{dbms(connection)}}
}
\value{
A DBI::Id that represents a qualified table and schema
}
\description{
Helper for working with compound schema
}
