% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.R
\name{intersect_cohorts}
\alias{intersect_cohorts}
\alias{intersectCohorts}
\title{Intersect all cohorts in a single cohort table}
\usage{
intersect_cohorts(x, cohort_definition_id = 1L)

intersectCohorts(x, cohort_definition_id = 1L)
}
\arguments{
\item{x}{A tbl reference to a cohort table}

\item{cohort_definition_id}{A number to use for the new cohort_definition_id

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
}
\value{
A lazy query that when executed will resolve to a new cohort table with
one cohort_definition_id resulting from the intersection of all cohorts in the original
cohort table
}
\description{
Intersect all cohorts in a single cohort table
}
