% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohortSet.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cohortCount}
\title{Get cohort counts from a generated_cohort_set object.}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{omopgenerics}{\code{\link[omopgenerics]{cohortCount}}}
}}

