% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSource.R, R/generateCohortSet.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dropTable}
\alias{cohortCount}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{omopgenerics}{\code{\link[omopgenerics]{cohortCount}}, \code{\link[omopgenerics]{dropTable}}}
}}

