% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSource.R, R/reexports-omopgenerics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{compute}
\alias{cohortCount}
\alias{settings}
\alias{attrition}
\alias{newCohortTable}
\alias{insertTable}
\alias{dropTable}
\alias{readSourceTable}
\alias{dropSourceTable}
\alias{listSourceTables}
\alias{bind}
\alias{cdmVersion}
\alias{cdmFromTables}
\alias{cohortCodelist}
\alias{cdmName}
\alias{recordCohortAttrition}
\alias{uniqueTableName}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{compute}}}

  \item{omopgenerics}{\code{\link[omopgenerics]{attrition}}, \code{\link[omopgenerics]{bind}}, \code{\link[omopgenerics]{cdmFromTables}}, \code{\link[omopgenerics]{cdmName}}, \code{\link[omopgenerics]{cdmVersion}}, \code{\link[omopgenerics]{cohortCodelist}}, \code{\link[omopgenerics]{cohortCount}}, \code{\link[omopgenerics]{dropSourceTable}}, \code{\link[omopgenerics]{dropTable}}, \code{\link[omopgenerics]{insertTable}}, \code{\link[omopgenerics]{listSourceTables}}, \code{\link[omopgenerics]{newCohortTable}}, \code{\link[omopgenerics]{readSourceTable}}, \code{\link[omopgenerics]{recordCohortAttrition}}, \code{\link[omopgenerics]{settings}}, \code{\link[omopgenerics]{uniqueTableName}}}
}}

