% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlrender.R
\name{translate_r}
\alias{translate_r}
\title{Translate SQL to target dialect
(Pure R implementation — kept as fallback)}
\usage{
translate_r(
  sql,
  targetDialect,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  oracleTempSchema = NULL
)
}
\arguments{
\item{sql}{SQL string.}

\item{targetDialect}{Target dialect.}

\item{tempEmulationSchema}{Temp schema for emulation.}

\item{oracleTempSchema}{Oracle temp schema.}
}
\description{
Translate SQL to target dialect
(Pure R implementation — kept as fallback)
}
\keyword{internal}
