% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{uniqueTableName}
\alias{uniqueTableName}
\alias{unique_table_name}
\title{Return a unique temp table name}
\usage{
uniqueTableName()

unique_table_name()
}
\value{
A string to be used for a temp table name
}
\description{
Return a unique temp table name
}
