% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm.R
\name{version}
\alias{version}
\title{Get the CDM version}
\usage{
version(cdm)
}
\arguments{
\item{cdm}{A cdm object}
}
\value{
"5.3" or "5.4"
}
\description{
Get the CDM version
}
\examples{
\dontrun{
library(CDMConnector)
con <- DBI::dbConnect(duckdb::duckdb(), eunomia_dir())
cdm <- cdm_from_con(con, cdm_tables = c(tbl_group("default"), "cdm_source"))
version(cdm)

DBI::dbDisconnect(con, shutdown = TRUE)
}
}
